refactor(session): move ordered-block helper to utils
Keep the server-owned content_blocks replay logic in a session-level utility so LinearTrajectory only owns when it is applied. The helper body and update timing remain unchanged.
