+++ miles/rollout/session/linear_trajectory.py
@@ -6,6 +6,7 @@ from typing import Any
 
 from miles.rollout.session.errors import MessageValidationError, SessionNotFoundError, TokenizationError
 from miles.rollout.session.types import SessionRecord
+from miles.rollout.session.utils import preserve_ordered_content_blocks
 from miles.utils.chat_template_utils import assert_messages_append_only_with_allowed_role, message_matches
 from miles.utils.chat_template_utils.tito_tokenizer import TITOTokenizer
 
@@ -15,34 +16,6 @@ logger = logging.getLogger(__name__)
 # TODO: hardcoded to 1 for now; if multi-step rollback is actually needed,
 #  raise this limit or make it configurable and remove the restriction.
 MAX_ASSISTANT_ROLLBACK_STEPS = 1
-
-
-def _preserve_ordered_content_blocks(
-    stored_messages: list[dict[str, Any]],
-    request_messages: list[dict[str, Any]],
-) -> list[dict[str, Any]]:
-    """Retain server-owned Inkling block order when clients replay history.
-
-    ``content_blocks`` is deliberately absent from the client response, so the
-    next request can only replay the compatible flattened OpenAI fields. Copy
-    the stored side channel solely across messages already proven equivalent by
-    ``message_matches``; never infer or synthesize it for new messages.
-    """
-    preserved = list(request_messages)
-    for index, request_message in enumerate(preserved):
-        if index >= len(stored_messages):
-            break
-        stored_message = stored_messages[index]
-        if (
-            stored_message.get("role") == "assistant"
-            and "content_blocks" in stored_message
-            and message_matches(stored_message, request_message)
-        ):
-            preserved[index] = {
-                **request_message,
-                "content_blocks": stored_message["content_blocks"],
-            }
-    return preserved
 
 
 def assert_pretokenized_prefix(
@@ -182,7 +155,7 @@ class LinearTrajectory:
             assistant_message=assistant_message,
         )
 
-        request_messages = _preserve_ordered_content_blocks(self.messages, request_messages)
+        request_messages = preserve_ordered_content_blocks(self.messages, request_messages)
         self.messages = request_messages + [assistant_message]
         self.trajectory_token_ids.append(all_token_ids)
         self.generated_checkpoint_message_ends.append(len(request_messages) + 1)
